requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
bleach==6.0.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.5
django-admin-geomap==1.6
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-constance==3.1.0
django-imagekit==4.1.0
django-picklefield==3.1
django-summernote==0.8.20.0
django-templated-mail==1.1.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
flake8==6.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
mccabe==0.7.0
numpy==1.26.0
oauthlib==3.2.2
openpyxl==3.1.0
packaging==23.1
pandas==2.1.0
pilkit==2.0
Pillow==9.5.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.3
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
